fix rounding error with Big Decimal and start and stop communicate on mainactivity resume and pause

+++ NxGnFirebaseModule/src/main/java/com/vismo/nxgnfirebasemodule/DashManager.kt
@@ -303,8 +303,8 @@ class DashManager @Inject constructor(
             updateTripOnFirestore(
                 MeterTripInFirestore(
                     tripId = tripId,
-                    total = total.roundTo(2),
-                    dashFee = fee.roundTo(2)
+                    total = total.roundTo(2).toDouble(),
+                    dashFee = fee.roundTo(2).toDouble()
                 )
             )
         }
